mv TimeBombDataStore
Move test-only DataStore implementation into test_utils.
