+++ core/src/utils/test_utils/data_store.rs
@@ -50,3 +50,104 @@ pub mod memory {
 pub fn null() -> NullDataStore {
     NullDataStore
 }
+
+pub fn time_bomb(t_minus: usize) -> internal::TimeBombDataStore {
+    internal::TimeBombDataStore::new(t_minus)
+}
+
+mod internal {
+    use crate::storage::{DataStore, MemoryDataStore};
+    use crate::world::thing::Thing;
+    use async_trait::async_trait;
+    use std::cell::RefCell;
+    use std::rc::Rc;
+    use uuid::Uuid;
+
+    pub struct TimeBombDataStore {
+        t_minus: Rc<RefCell<usize>>,
+        data_store: MemoryDataStore,
+    }
+
+    impl TimeBombDataStore {
+        pub fn new(t_minus: usize) -> Self {
+            Self {
+                t_minus: Rc::new(t_minus.into()),
+                data_store: MemoryDataStore::default(),
+            }
+        }
+
+        fn tick(&self) -> Result<(), ()> {
+            if *self.t_minus.borrow() == 0 {
+                Err(())
+            } else {
+                self.t_minus.replace_with(|&mut i| i - 1);
+                Ok(())
+            }
+        }
+    }
+
+    #[async_trait(?Send)]
+    impl DataStore for TimeBombDataStore {
+        async fn health_check(&self) -> Result<(), ()> {
+            if *self.t_minus.borrow() == 0 {
+                Err(())
+            } else {
+                Ok(())
+            }
+        }
+
+        async fn delete_thing_by_uuid(&mut self, uuid: &Uuid) -> Result<(), ()> {
+            self.tick()?;
+            self.data_store.delete_thing_by_uuid(uuid).await
+        }
+
+        async fn edit_thing(&mut self, thing: &Thing) -> Result<(), ()> {
+            self.tick()?;
+            self.data_store.edit_thing(thing).await
+        }
+
+        async fn get_all_the_things(&self) -> Result<Vec<Thing>, ()> {
+            self.tick()?;
+            self.data_store.get_all_the_things().await
+        }
+
+        async fn get_thing_by_uuid(&self, uuid: &Uuid) -> Result<Option<Thing>, ()> {
+            self.tick()?;
+            self.data_store.get_thing_by_uuid(uuid).await
+        }
+
+        async fn get_thing_by_name(&self, name: &str) -> Result<Option<Thing>, ()> {
+            self.tick()?;
+            self.data_store.get_thing_by_name(name).await
+        }
+
+        async fn get_things_by_name_start(
+            &self,
+            name: &str,
+            limit: Option<usize>,
+        ) -> Result<Vec<Thing>, ()> {
+            self.tick()?;
+            self.data_store.get_things_by_name_start(name, limit).await
+        }
+
+        async fn save_thing(&mut self, thing: &Thing) -> Result<(), ()> {
+            self.tick()?;
+            self.data_store.save_thing(thing).await
+        }
+
+        async fn set_value(&mut self, key: &str, value: &str) -> Result<(), ()> {
+            self.tick()?;
+            self.data_store.set_value(key, value).await
+        }
+
+        async fn get_value(&self, key: &str) -> Result<Option<String>, ()> {
+            self.tick()?;
+            self.data_store.get_value(key).await
+        }
+
+        async fn delete_value(&mut self, key: &str) -> Result<(), ()> {
+            self.tick()?;
+            self.data_store.delete_value(key).await
+        }
+    }
+}
